@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2015 the original author or authors.
+ * Copyright the original author or authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -12,15 +12,14 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 package de.schildbach.wallet.offline;
 
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
-import java.util.concurrent.atomic.AtomicBoolean;
+import android.bluetooth.BluetoothAdapter;
+import android.bluetooth.BluetoothServerSocket;
+import android.bluetooth.BluetoothSocket;
 
 import org.bitcoin.protocols.payments.Protos;
 import org.bitcoin.protocols.payments.Protos.PaymentACK;
@@ -30,12 +29,13 @@ import org.bitcoinj.protocols.payments.PaymentProtocol;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+import java.io.IOException;
+import java.util.concurrent.atomic.AtomicBoolean;
+
 import de.schildbach.wallet.Constants;
 import de.schildbach.wallet.util.Bluetooth;
-
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothServerSocket;
-import android.bluetooth.BluetoothSocket;
 
 /**
  * @author Shahar Livne
@@ -62,18 +62,11 @@ public abstract class AcceptBluetoothThread extends Thread {
             org.bitcoinj.core.Context.propagate(Constants.CONTEXT);
 
             while (running.get()) {
-                BluetoothSocket socket = null;
-                DataInputStream is = null;
-                DataOutputStream os = null;
-
-                try {
-                    // start a blocking call, and return only on success or exception
-                    socket = listeningSocket.accept();
-
+                try ( // start a blocking call, and return only on success or exception
+                        final BluetoothSocket socket = listeningSocket.accept();
+                        final DataInputStream is = new DataInputStream(socket.getInputStream());
+                        final DataOutputStream os = new DataOutputStream(socket.getOutputStream())) {
                     log.info("accepted classic bluetooth connection");
-
-                    is = new DataInputStream(socket.getInputStream());
-                    os = new DataOutputStream(socket.getOutputStream());
 
                     boolean ack = true;
 
@@ -98,30 +91,6 @@ public abstract class AcceptBluetoothThread extends Thread {
                     os.writeBoolean(ack);
                 } catch (final IOException x) {
                     log.info("exception in bluetooth accept loop", x);
-                } finally {
-                    if (os != null) {
-                        try {
-                            os.close();
-                        } catch (final IOException x) {
-                            // swallow
-                        }
-                    }
-
-                    if (is != null) {
-                        try {
-                            is.close();
-                        } catch (final IOException x) {
-                            // swallow
-                        }
-                    }
-
-                    if (socket != null) {
-                        try {
-                            socket.close();
-                        } catch (final IOException x) {
-                            // swallow
-                        }
-                    }
                 }
             }
         }
@@ -138,18 +107,11 @@ public abstract class AcceptBluetoothThread extends Thread {
             org.bitcoinj.core.Context.propagate(Constants.CONTEXT);
 
             while (running.get()) {
-                BluetoothSocket socket = null;
-                DataInputStream is = null;
-                DataOutputStream os = null;
-
-                try {
-                    // start a blocking call, and return only on success or exception
-                    socket = listeningSocket.accept();
-
+                try ( // start a blocking call, and return only on success or exception
+                        final BluetoothSocket socket = listeningSocket.accept();
+                        final DataInputStream is = new DataInputStream(socket.getInputStream());
+                        final DataOutputStream os = new DataOutputStream(socket.getOutputStream())) {
                     log.info("accepted payment protocol bluetooth connection");
-
-                    is = new DataInputStream(socket.getInputStream());
-                    os = new DataOutputStream(socket.getOutputStream());
 
                     boolean ack = true;
 
@@ -171,30 +133,6 @@ public abstract class AcceptBluetoothThread extends Thread {
                     paymentAck.writeDelimitedTo(os);
                 } catch (final IOException x) {
                     log.info("exception in bluetooth accept loop", x);
-                } finally {
-                    if (os != null) {
-                        try {
-                            os.close();
-                        } catch (final IOException x) {
-                            // swallow
-                        }
-                    }
-
-                    if (is != null) {
-                        try {
-                            is.close();
-                        } catch (final IOException x) {
-                            // swallow
-                        }
-                    }
-
-                    if (socket != null) {
-                        try {
-                            socket.close();
-                        } catch (final IOException x) {
-                            // swallow
-                        }
-                    }
                 }
             }
         }
